pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kalshi-cricket-tracker"
version = "0.1.0"
description = "Cricket odds proxy + Kalshi paper-trading research assistant"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.1",
  "numpy>=1.26",
  "requests>=2.31",
  "pyyaml>=6.0",
  "typer>=0.12",
  "rich>=13.7",
  "pydantic>=2.6",
  "streamlit>=1.32",
  "plotly>=5.18"
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=4.1"]

[project.scripts]
kct = "kalshi_cricket_tracker.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
